# Optimize finding singular target elements