Tested (somewhat unscientifically) using the following markup nested several levels deep:
```html
<div data-controller="x" data-target="x.a">
  <span>
    <span data-target="x.b"></span>
    <span data-target="x.b"></span>
  </span>

  <span data-target="x.c"></span>
  <span data-target="x.c"></span>

  <div data-controller="x" data-target="x.a">
    …
  </div>
</div>
```

Before:
```
this.aTarget 0.140ms
this.bTarget 0.230ms
this.cTarget 0.160ms
```

After:
```
this.aTarget 0.035ms
this.bTarget 0.140ms
this.cTarget 0.110ms
```