@sstephenson 2019-08-15
The code itself looks fine 👍 I’ll leave it up to you as far as merging, though.

I’d prefer not to make micro-optimizations like this without some sort of corpus we can benchmark against. As you’ve noted, the test case here is unscientific and not really representative of an actual DOM tree, and AFAIK we don’t have any reason to think the current implementation will be a performance problem. We also wouldn’t have any way to know if we were to regress in the future.